LR Minibatch Study consolidated and definition updated
